break up triangle stats container component into triangle stats and triangle image
