+++ web/src/stylesheets/triangle-stats.css
@@ -1,0 +1,10 @@
+.triangle-stats-list {
+  flex: 1;
+  height: 90%;
+  display: flex;
+  flex-direction: column;
+  text-align: start;
+  /* border-radius: 20px; */
+  margin-left: 2.5%;
+  overflow: auto;
+}
